% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eotsfilter.R
\name{addMod13Q1quality}
\alias{addMod13Q1quality}
\title{DEPRECATED (moved to scidbutil) - Add quality description to MOD13Q1}
\usage{
addMod13Q1quality(x)
}
\arguments{
\item{x}{A data frame containing the column c("quality")}
}
\value{
A data frame with additional columns.
}
\description{
Add readable quality data to a MOD13Q1 time series according to its data quality
}
