% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eotsfilter.R
\name{doublelogistic}
\alias{doublelogistic}
\title{Warpper of the double logistic fit}
\usage{
doublelogistic(sample)
}
\arguments{
\item{sample}{A list. A SAMPLE object processed using jsonlite::fromJSON}
}
\value{
A list representing A JSON object of type SAMPLE
}
\description{
Warpper of the double logistic fit
}
