% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{fill}
\alias{fill}
\title{Fill in the gaps of a time series}
\usage{
fill(type, sample)
}
\arguments{
\item{type}{A numeric. The type of fill}

\item{sample}{A list. A SAMPLE object processed using jsonlite::fromJSON}
}
\value{
A list representing A JSON object of type SAMPLE
}
\description{
Fill in the gaps of a time series
}
