% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier.R
\name{fourier}
\alias{fourier}
\title{Warpper of the Fourier filter}
\usage{
fourier(nfreq, sample)
}
\arguments{
\item{nfreq}{A numeric. Number of low frequencies to keep}

\item{sample}{A list. A SAMPLE object processed using jsonlite::fromJSON}
}
\value{
A list representing A JSON object of type SAMPLE
}
\description{
Warpper of the Fourier filter
}
