% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eotsfilter.R
\name{mod13q1quality2codes}
\alias{mod13q1quality2codes}
\title{Cast MOD13Q1 quality data to codes}
\usage{
mod13q1quality2codes(x)
}
\arguments{
\item{x}{A data frame containing the column c("quality")}
}
\value{
A data frame with additional columns.
}
\description{
Cast MOD13Q1 quality data to codes
}
\details{
See TABLE 2: MOD13Q1 VI Quality at https://lpdaac.usgs.gov/dataset_discovery/modis/modis_products_table/mod13q1
}
