% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples1}
\alias{samples1}
\title{Time series of deforested points in the Amazon forest}
\format{A list of 337 elements.
\describe{
  \item{sample}{A data frame with 1 row and 14 variables. It contains the
  sample point data.
    \describe{
      \item{id}{sample id}
      \item{longitude}{sample WGS84 longitude}
      \item{latitude}{sample WGS84 latitude}
      \item{start_date}{sample start date}
      \item{end_date}{sample start date}
      \item{label}{class to which the time time series mathches iduring the period given by the start and end dates}
      \item{linkcolumn}{DETER link column}
      \item{class_name}{DETER class name}
      \item{scene_id}{DETER scene id}
      \item{areameters}{DETER areameters}
      \item{view_date}{DETER date when deforestation was detected}
      \item{julday}{DETER day of the year of deforestation detection}
      \item{col_id}{SciDB column ID in the MODIS MOD13Q1 array}
      \item{row_id}{SciDB row ID in the MODIS MOD13Q1 array}
    }
  }
  \item{time_series}{A data frame with 363 rows and 15 variables. It contains
  the time series of MOD13Q1 data of the sample point.
    \describe{
      \item{col_id}{SciDB column ID in the MODIS MOD13Q1 array}
      \item{row_id}{SciDB row ID in the MODIS MOD13Q1 array}
      \item{time_id}{SciDB time ID in the MODIS MOD13Q1 array}
      \item{ndvi}{MOD13Q1 Normalized Difference Vegetation Index}
      \item{evi}{MOD13Q1 Enhanced Vegetation Index}
      \item{quality}{MOD13Q1 quality detailed}
      \item{red}{MOD13Q1 red reflectance (Band 1)}
      \item{nir}{MOD13Q1 Near Infra Red reflectance (Band 2)}
      \item{blue}{MOD13Q1 blue reflectance (Band 3)}
      \item{mir}{MOD13Q1 Middle Infra Red reflectance (Band 7)}
      \item{view_zenith}{MOD13Q1 view zenith angle}
      \item{sun_zenith}{MOD13Q1 sun zenith angle}
      \item{relative_azimuth}{MOD13Q1 relative azimuth angle}
      \item{day_of_year}{MOD13Q1 julian day of year}
      \item{reliability}{MOD13Q1 pixel reliability summary QA}
    }
  }
}}
\source{
\url{https://github.com/e-sensing/sits/blob/master/inst/extdata/samples/deforestation_points_226_64.csv}
}
\usage{
samples1
}
\description{
A dataset containing samples of deforestation in the Amazon forest. These
samples were collected by Rodrigo Anzolin Begotti who works for the e-sensing
project. The original CSV files are spatially joined to DETER data to get the
reported deforestation date of each sample. The data is also joined to
MOD13Q1 data to get the time series of vegetation indexes.
}
\keyword{datasets}
