% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{siteA}
\alias{siteA}
\title{Time series of vegetation indexes of Manicorao, Brazil}
\format{A data frame with 1092 rows and 15 variables:
\describe{
   \item{col_id}{SciDB column ID in the MODIS MOD13Q1 array}
   \item{row_id}{SciDB row ID in the MODIS MOD13Q1 array}
   \item{time_id}{SciDB time ID in the MODIS MOD13Q1 array}
   \item{ndvi}{MOD13Q1 Normalized Difference Vegetation Index}
   \item{evi}{MOD13Q1 Enhanced Vegetation Index}
   \item{quality}{MOD13Q1 quality detailed}
   \item{red}{MOD13Q1 red reflectance (Band 1)}
   \item{nir}{MOD13Q1 Near Infra Red reflectance (Band 2)}
   \item{blue}{MOD13Q1 blue reflectance (Band 3)}
   \item{mir}{MOD13Q1 Middle Infra Red reflectance (Band 7)}
   \item{view_zenith}{MOD13Q1 view zenith angle}
   \item{sun_zenith}{MOD13Q1 sun zenith angle}
   \item{relative_azimuth}{MOD13Q1 relative azimuth angle}
   \item{day_of_year}{MOD13Q1 julian day of year}
   \item{reliability}{MOD13Q1 pixel reliability summary QA}
}}
\source{
\url{http://www.dpi.inpe.br/gilberto/teses/msc_thesis_christopher_stephan.pdf}
}
\usage{
siteA
}
\description{
A dataset of MODIS MOD13Q1 data used in Christopher Stephan' thesis
\emph{Automating Near Real-Time Deforestation Monitoring With Satellite Image
Time Series.}. These belongs to siteA
}
\keyword{datasets}
