% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eotsfilter.R
\name{sts}
\alias{sts}
\title{Warpper of the Fit Structural Time Series}
\usage{
sts(sample)
}
\arguments{
\item{sample}{A list. A SAMPLE object processed using jsonlite::fromJSON}
}
\value{
A list representing A JSON object of type SAMPLE
}
\description{
Warpper of the Fit Structural Time Series
}
