% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitaker.R
\name{whitaker2}
\alias{whitaker2}
\title{Wrapper of the Weighted Whittaker smoothing with a second order finite difference penalty}
\usage{
whitaker2(lambda, sample)
}
\arguments{
\item{lambda}{A numeric. The smoothing parameter}

\item{sample}{A list. A SAMPLE object processed using jsonlite::fromJSON}
}
\value{
A list representing A JSON object of type SAMPLE
}
\description{
Wrapper of the Weighted Whittaker smoothing with a second order finite difference penalty
}
